fix `perform': select(): Operation now in progress` in ruby 1.8.7
This appears to be a result of misplaced `#if` and `#endif` caused the
default fallback to never be executed and `rc` would remain with
initialized value (`-1`) and raise an exception:

```
      switch(rc) {
      case -1:
        if(errno != EINTR) {
          rb_raise(rb_eRuntimeError, "select(): %s", strerror(errno));
          break;
        }
```

This was the original flow with indentation:

```
 #if defined(HAVE_RB_THREAD_BLOCKING_REGION) || defined(HAVE_RB_THREAD_CALL_WITHOUT_GVL)
   fdset_args.maxfd = maxfd+1;
   fdset_args.fdread = &fdread;
   fdset_args.fdwrite = &fdwrite;
   fdset_args.fdexcep = &fdexcep;
   fdset_args.tv = &tv;
   #ifdef HAVE_RB_THREAD_CALL_WITHOUT_GVL
      rc = (int)(VALUE) rb_thread_call_without_gvl((void *(*)(void *))curb_select, &fdset_args, RUBY_UBF_IO, 0);
   #elif HAVE_RB_THREAD_BLOCKING_REGION
      rc = rb_thread_blocking_region(curb_select, &fdset_args, RUBY_UBF_IO, 0);
   #elif HAVE_RB_THREAD_FD_SELECT
      rc = rb_thread_fd_select(maxfd+1, &fdread, &fdwrite, &fdexcep, &tv);
   #else
      rc = rb_thread_select(maxfd+1, &fdread, &fdwrite, &fdexcep, &tv);
   #endif
 #endif
```

Ruby 1.8 does not define `HAVE_RB_THREAD_BLOCKING_REGION` nor
`HAVE_RB_THREAD_CALL_WITHOUT_GVL` and no code is executed because the
`else` clausule applies to inner `ifdef`.

New flow, after my changes with indentation:

```
 #if defined(HAVE_RB_THREAD_BLOCKING_REGION) || defined(HAVE_RB_THREAD_CALL_WITHOUT_GVL)
   fdset_args.maxfd = maxfd+1;
   fdset_args.fdread = &fdread;
   fdset_args.fdwrite = &fdwrite;
   fdset_args.fdexcep = &fdexcep;
   fdset_args.tv = &tv;
 #endif

 #ifdef HAVE_RB_THREAD_CALL_WITHOUT_GVL
    rc = (int)(VALUE) rb_thread_call_without_gvl((void *(*)(void *))curb_select, &fdset_args, RUBY_UBF_IO, 0);
 #elif HAVE_RB_THREAD_BLOCKING_REGION
    rc = rb_thread_blocking_region(curb_select, &fdset_args, RUBY_UBF_IO, 0);
 #elif HAVE_RB_THREAD_FD_SELECT
    rc = rb_thread_fd_select(maxfd+1, &fdread, &fdwrite, &fdexcep, &tv);
 #else
    rc = rb_thread_select(maxfd+1, &fdread, &fdwrite, &fdexcep, &tv);
 #endif
```

Here even when all the features are unavailable we have a fallback to
`rc = rb_thread_select(maxfd+1, &fdread, &fdwrite, &fdexcep, &tv);`.

This commit also changes `rb_thread_wait_for` to pass timeval
explicityly rather than rely on `rb_time_timeval`. There must have been
a change in ruby 1.9 because the `rb_time_timeval` does not work in
1.8.7.

```
-        rb_thread_wait_for(rb_time_timeval(DBL2NUM(0.1)));
+        rb_thread_wait_for(tv_100ms);
```