fix `bundle install --deployment` (#350)
* fix `bundle install --deployment`

Travis runs `bundle install --deployment` and it fails because the
Gemfile.lock is not up to date.

`curb` version has been updated, but `bundle install` has not been
executed. Bundler is expected to fail with `--deployment` flag if
Gemfile.lock is not up to date.

This commit also removes deprecated fields from the gemspec file.

* fix cookie decoding bug in test helper

`decode_www_form` would throw an exception because input was invalid
(not a properly encoded form). We're passing single component here and
`decode_www_form_component` should be used for decoding.