do not append `?` to urls if there are no query parameters (#345)
* do not append `?` to urls if there are no query parameters

Given an URL without query params Curb would add `?` at the end. This
has been introduced in commit 32fa6d78968c3b63e2a54a2c326efb577db04043.

This happens because we set `uri.query` to empty string and URI will
treat this as a valid query and it will add `?`. This does not happen
when the query is set to nil.

I have decide to early return if extra parameters are not passed. It
will work whatever there were query params in the original url or not
because we don't change `#query` and it's set correctly.

Added unit tests for all use cases.

The empty `?` is meangingless for the browser/server, but it can be
significant for caching strategies. I use `curb` to purge Varnish
cache for individual urls and addition of `?` breaks this because
the purged URL is different from what I expected.

* refactor Curl.urlalize

This greatly simplifies the logic by removing branching. The `uri.query`
is always computed the same way, no matter if there are parameters or
not.

The tests have been added in the previous commit and this change is not
expected to change the method behavior.

* Curl.urlalize can handle nil again

The early return added before would cause a crash if `nil` was passed to
the method. It fixes this regression and includes a test case.