fix timeout and timeout_ms to handle floats and negative numbers
The setter used `CURB_IMMED_SETTER` macro and it casts value to `long`
and loses precision if float was passed. C will round down as part of
the conversion and if it converts to 0 then the IMMED_GETTER will return
nil.

This results in a confusing output described in #355.

This commit deperecates `CURLOPT_TIMEOUT` option and the code will rely
exclusively on `timeout_ms` (and `CURLOPT_TIMEOUT_MS` accordingly).

There are minor user-facing changes:

* timeout and timeout_ms will not return nil anymore
* change of timeout will now also change timeout_ms respectively
* timeout will return numeric rather than fixnum

Otherwise everything should be working as expected.