fixes to make curb work on ruby 1.8
Before Ruby 1.8 was removed from Travis it was failing tests related to
outdated lockfile as a result of new curb release.

I had to prepare a Docker environment with Ruby 1.8 to update lockfile
correctly. On top of that this P/R does:

* Rakefile & friends general ruby syntax fixes
  - splat operator is not available in Ruby 1.8
  - new and compact hash syntax is not available in Ruby 1.8
  - popen3 does not yield wait thread in Ruby 1.8
* macro DBL2NUM has been defined
  - not available in Ruby 1.8

There are no official images for Ruby 1.8 and I had to chose
something... Phusion images used to be reliable.

Ruby 1.8 image comes without libcurl development libraries and I decided
to override an entrypoint and install it on demand. It's slow, but it is
enough to check if curb still works on Ruby 1.8.