skip `test_connection_keepalive` if libcurl < 7.22.0
Maybe it's a bug in libcurl and curb shouldn't attempt to backport
patches.

This also includes minor changes to test implementation. `ss` should be
available on more platforms than `lsof`.

The endpoint has been changed to `TestServlet.url` and we don't need
external connectivity anymore. The `$TEST_URL` doesn't work here because
it's a `file://` type of URL.