add support for `CURLOPT_STDERR`
This allows to capture verbose output without hijacking global `STDERR`.
`libcurl` requires a raw `FILE` pointer so it works only with IO objects
in Ruby. String or StringIO won't do.

This is as simple as I could make it and it will either work or explode
with a meaningful exception ("wrong argument type Tempfile (expected File)").

Tested on rubies 1.8 up to 2.6.