fix enabling haproxy protocol
The following code does not have the intended behavior

  curl.setopt(Curl::CURLOPT_HAPROXYPROTOCOL, 1)

The expected header does not get set. This looks like a small typo.

Instead of passing the truthy HAVE_CURLOPT_HAPROXYPROTOCOL we should
be passing CURLOPT_HAPROXYPROTOCOL to curl_easy_setopt.

Validated by rebuilding the gem with the changes and re-testing.